split signed and fixedpoint implementations
