+++ src/peakrdl_regblock/udps/__init__.py
@@ -1,7 +1,8 @@
 from .rw_buffering import BufferWrites, WBufferTrigger
 from .rw_buffering import BufferReads, RBufferTrigger
 from .extended_swacc import ReadSwacc, WriteSwacc
-from .fixedpoint import IntWidth, FracWidth, IsSigned
+from .fixedpoint import IntWidth, FracWidth
+from .signed import IsSigned
 
 ALL_UDPS = [
     BufferWrites,
