+++ src/peakrdl_regblock/udps/fixedpoint.py
@@ -67,29 +67,3 @@ class FracWidth(FixedpointWidth):
         else:
             # not a fixed-point number
             return None
-
-
-class IsSigned(UDPDefinition):
-    name = "is_signed"
-    valid_components = {Field}
-    valid_type = bool
-    default_assignment = True
-
-    def validate(self, node: "Node", value: Any) -> None:
-        # "counter" fields can not be signed
-        if value and node.get_property("counter"):
-            self.msg.error(
-                "The property is_signed=true is not supported for counter fields.",
-                node.inst.property_src_ref["is_signed"]
-            )
-
-        # incompatible with "encode" fields
-        if value and node.get_property("encode") is not None:
-            self.msg.error(
-                "The property is_signed=true is not supported for fields encoded as an enum.",
-                node.inst.property_src_ref["is_signed"]
-            )
-
-    def get_unassigned_default(self, node: "Node") -> Any:
-        # unsigned if not specified
-        return False
